+++ demis/applications/chart-notification_gateway.tf
@@ -25,25 +25,27 @@ module "notification_gateway" {
 
   # Pass the values for the chart
   application_values = templatefile(local.gateway_template_app, {
-    image_pull_secrets = var.pull_secrets,
-    repository         = var.docker_registry,
-    namespace          = var.target_namespace,
-    debug_enable       = var.debug_enabled,
-    istio_enable       = var.istio_enabled,
-    portal_hostname    = var.portal_hostname,
-    meldung_hostname   = var.meldung_hostname,
-    core_hostname      = var.core_hostname,
-    issuer_hostname    = var.auth_hostname,
-    context_path       = var.context_path,
-    feature_flags      = try(var.feature_flags[local.gateway_name], {}),
-    config_options     = try(var.config_options[local.gateway_name], {}),
-    replica_count      = local.gateway_replicas,
-    resource_block     = local.gateway_resource_block
+    image_pull_secrets             = var.pull_secrets,
+    repository                     = var.docker_registry,
+    namespace                      = var.target_namespace,
+    debug_enable                   = var.debug_enabled,
+    istio_enable                   = var.istio_enabled,
+    portal_hostname                = var.portal_hostname,
+    meldung_hostname               = var.meldung_hostname,
+    core_hostname                  = var.core_hostname,
+    issuer_hostname                = var.auth_hostname,
+    context_path                   = var.context_path,
+    feature_flags                  = try(var.feature_flags[local.gateway_name], {}),
+    config_options                 = try(var.config_options[local.gateway_name], {}),
+    replica_count                  = local.gateway_replicas,
+    resource_block                 = local.gateway_resource_block,
+    feature_flag_new_api_endpoints = try(var.feature_flags[local.gateway_name].FEATURE_FLAG_NEW_API_ENDPOINTS, false)
   })
   istio_values = templatefile(local.gateway_template_istio, {
-    namespace        = var.target_namespace,
-    context_path     = var.context_path,
-    cluster_gateway  = var.cluster_gateway,
-    portal_hostnames = local.frontend_hostnames
+    namespace                      = var.target_namespace,
+    context_path                   = var.context_path,
+    cluster_gateway                = var.cluster_gateway,
+    portal_hostnames               = local.frontend_hostnames
+    feature_flag_new_api_endpoints = try(var.feature_flags[local.gateway_name].FEATURE_FLAG_NEW_API_ENDPOINTS, false)
   })
 }