+++ demis/applications/chart-report_processing_service.tf
@@ -24,21 +24,26 @@ module "report_processing_service" {
 
   # Pass the values for the chart
   application_values = templatefile(local.rps_template_app, {
-    image_pull_secrets         = var.pull_secrets,
-    repository                 = var.docker_registry,
-    namespace                  = var.target_namespace,
-    debug_enable               = var.debug_enabled,
-    istio_enable               = var.istio_enabled,
-    keycloak_internal_hostname = var.keycloak_internal_hostname,
-    feature_flags              = try(var.feature_flags[local.rps_name], {}),
-    config_options             = try(var.config_options[local.rps_name], {}),
-    replica_count              = local.rps_replicas,
-    resource_block             = local.rps_resource_block
+    image_pull_secrets             = var.pull_secrets,
+    repository                     = var.docker_registry,
+    namespace                      = var.target_namespace,
+    debug_enable                   = var.debug_enabled,
+    istio_enable                   = var.istio_enabled,
+    keycloak_internal_hostname     = var.keycloak_internal_hostname,
+    feature_flags                  = try(var.feature_flags[local.rps_name], {}),
+    config_options                 = try(var.config_options[local.rps_name], {}),
+    replica_count                  = local.rps_replicas,
+    resource_block                 = local.rps_resource_block,
+    feature_flag_new_api_endpoints = try(var.feature_flags[local.rps_name].FEATURE_FLAG_NEW_API_ENDPOINTS, false)
   })
   istio_values = templatefile(local.rps_template_istio, {
-    namespace       = var.target_namespace,
-    context_path    = var.context_path,
-    cluster_gateway = var.cluster_gateway,
-    core_hostname   = var.core_hostname
+    namespace                      = var.target_namespace,
+    context_path                   = var.context_path,
+    cluster_gateway                = var.cluster_gateway,
+    core_hostname                  = var.core_hostname
+    support_fhir_api_versions      = var.profile_provisioning_mode_vs_core != null && var.profile_provisioning_mode_vs_core != "dedicated"
+    fhir_api_versions              = module.validation_service_core_metadata.current_profile_versions
+    feature_flag_new_api_endpoints = try(var.feature_flags[local.rps_name].FEATURE_FLAG_NEW_API_ENDPOINTS, false)
+
   })
 }