+++ demis/variables_applications.tf
@@ -50,3 +50,36 @@ variable "fhir_storage_purger_cron_schedule" {
     error_message = "The FHIR storage purger cron schedule must be defined"
   }
 }
+
+variable "profile_provisioning_mode_vs_core" {
+  description = "Provisioning mode for the FHIR Profiles services. Allowed values are: dedicated, distributed, combined"
+  type        = string
+  nullable    = true
+  default     = null
+  validation {
+    condition     = var.profile_provisioning_mode_vs_core == null || try(contains(["dedicated", "distributed", "combined"], var.profile_provisioning_mode_vs_core), false)
+    error_message = "The provisioning mode must be one of the following: dedicated, distributed, combined"
+  }
+}
+
+variable "profile_provisioning_mode_vs_igs" {
+  description = "Provisioning mode for the FHIR Profiles services. Allowed values are: dedicated, distributed, combined"
+  type        = string
+  nullable    = true
+  default     = null
+  validation {
+    condition     = var.profile_provisioning_mode_vs_igs == null || try(contains(["dedicated", "distributed", "combined"], var.profile_provisioning_mode_vs_igs), false)
+    error_message = "The provisioning mode must be one of the following: dedicated, distributed, combined"
+  }
+}
+
+variable "profile_provisioning_mode_vs_ars" {
+  description = "Provisioning mode for the FHIR Profiles services. Allowed values are: dedicated, distributed, combined"
+  type        = string
+  nullable    = true
+  default     = null
+  validation {
+    condition     = var.profile_provisioning_mode_vs_ars == null || try(contains(["dedicated", "distributed", "combined"], var.profile_provisioning_mode_vs_ars), false)
+    error_message = "The provisioning mode must be one of the following: dedicated, distributed, combined"
+  }
+}